+++ lib/objectstore/objectid/objectid.go
@@ -1,4 +1,4 @@
-package objectstore
+package objectid
 
 // ObjectID contains the string representation of an objects identifier.
 type ObjectID string
